Move LB config to a service_config module to avoid circulr deps with resolver API
