+++ grpc/src/client/service_config.rs
@@ -15,8 +15,34 @@
  * limitations under the License.
  *
  */
+use std::{any::Any, error::Error, sync::Arc};
 
 /// An in-memory representation of a service config, usually provided to gRPC as
 /// a JSON object.
 #[derive(Debug, Default, Clone)]
 pub struct ServiceConfig;
+
+/// A convenience wrapper for an LB policy's configuration object.
+#[derive(Debug)]
+pub struct LbConfig {
+    config: Arc<dyn Any + Send + Sync>,
+}
+
+impl LbConfig {
+    /// Create a new LbConfig wrapper containing the provided config.
+    pub fn new<T: 'static + Send + Sync>(config: T) -> Self {
+        LbConfig {
+            config: Arc::new(config),
+        }
+    }
+
+    /// Convenience method to extract the LB policy's configuration object.
+    pub fn convert_to<T: 'static + Send + Sync>(
+        &self,
+    ) -> Result<Arc<T>, Box<dyn Error + Send + Sync>> {
+        match self.config.clone().downcast::<T>() {
+            Ok(c) => Ok(c),
+            Err(e) => Err("failed to downcast to config type".into()),
+        }
+    }
+}
